# backend/static/openapi.yaml
openapi: 3.0.3
info:
  title: Gingerly API
  description: Email notification platform — send individual and bulk emails to contacts and groups.
  version: 1.0.0
servers:
  - url: http://localhost:5001/api
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Contact:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
...
    Group:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        contact_count:
          type: integer
        contacts:
          type: array
          items:
            $ref: '#/components/schemas/Contact'
        created_at:
          type: string
          format: date-time
    GroupEmailStatus:
      type: object
      properties:
        total:
          type: integer
        sent:
          type: integer
        pending:
          type: integer
        failed_count:
          type: integer
        failed_contacts:
          type: array
          items:
            type: object
            properties:
              contact_id:
                type: integer
              contact:
                $ref: '#/components/schemas/Contact'
...
security:
  - bearerAuth: []
paths:
...
  /groups:
    get:
      tags: [Groups]
      summary: List all groups (Gold plan only)
      responses:
        '200':
          description: List of groups
          content:
            application/json:
              schema:
                type: object
                properties:
                  groups:
                    type: array
                    items:
                      $ref: '#/components/schemas/Group'
        '403':
          description: Gold plan required
    post:
      tags: [Groups]
      summary: Create a new group (Gold plan only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                  example: Marketing Team
      responses:
        '201':
          description: Group created
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  group:
                    $ref: '#/components/schemas/Group'
        '403':
          description: Gold plan required
  /groups/{id}:
    delete:
      tags: [Groups]
      summary: Delete a group (Gold plan only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Group deleted
        '403':
          description: Gold plan required
        '404':
          description: Group not found
  /groups/{id}/contacts:
    post:
      tags: [Groups]
      summary: Add a contact to a group (Gold plan only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [contact_id]
              properties:
                contact_id:
                  type: integer
                  example: 1
      responses:
        '200':
          description: Contact added to group
        '403':
          description: Gold plan required
        '404':
          description: Group or contact not found
        '409':
          description: Contact already in group
  /groups/{id}/contacts/{contact_id}:
    delete:
      tags: [Groups]
      summary: Remove a contact from a group (Gold plan only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: contact_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Contact removed from group
        '403':
          description: Gold plan required
        '404':
          description: Group, contact, or membership not found
  /groups/{id}/emails:
    get:
      tags: [Groups]
      summary: List all email campaigns sent to a group (Gold plan only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: List of group emails with delivery status
    post:
      tags: [Groups]
      summary: Send an email to all contacts in a group (Gold plan only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [subject, body]
              properties:
                subject:
                  type: string
                  example: Team Update
                body:
                  type: string
                  example: Please review the latest report.
      responses:
        '201':
          description: Group email sent with delivery breakdown
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  group_email:
                    type: object
                    properties:
                      id:
                        type: integer
                      subject:
                        type: string
                      status:
                        $ref: '#/components/schemas/GroupEmailStatus'
        '400':
          description: Group has no contacts
        '403':
          description: Gold plan required
  /groups/{id}/emails/{email_id}/status:
    get:
      tags: [Groups]
      summary: Get delivery status of a group email (Gold plan only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: email_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Delivery status breakdown
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    $ref: '#/components/schemas/GroupEmailStatus'
        '403':
          description: Gold plan required
        '404':
          description: Group email not found
...